Execute operators for same event in parallel and entity_processed as a function instead of an operator

+++ lib/event_streamex/operators/queue.ex
@@ -5,11 +5,11 @@ defmodule EventStreamex.Operators.Queue do
   Every event received is dispatched to operators (`EventStreamex.Operators.Operator`),
   and executed sequentially to keep ordenancing.
 
-  Every time an event is received, a task is added to the queue of events for each
+  Every time an event is received, a task is added to the queue of events with each
   operator listening for this event.
-  And everytime a an operator task is finished, it is removed from the queue.
+  And everytime a an operator task is finished, it is update in the queue, until it is deleted completely when all operators have finished their task.
 
-  A queue item is a tuple containing the operator module and the event to process.
+  A queue item is a tuple containing the operator modules with their completion status and the event to process.
 
   The queue is handled internally.
   You should not use it directly.
@@ -64,9 +64,11 @@ defmodule EventStreamex.Operators.Queue do
   Adds another item to the queue
   """
   @doc since: "1.0.0"
-  def enqueue(module, event) do
+  def enqueue([], _event), do: :ok
+
+  def enqueue(modules, event) do
     Agent.update(__MODULE__, fn queue ->
-      new_item = {UUID.uuid4(), {module, event}}
+      new_item = {UUID.uuid4(), {modules |> Map.from_keys(false), event}}
       {:ok, _res} = QueueStorageAdapter.add_item(new_item)
 
       :queue.in(new_item, queue)
@@ -74,18 +76,31 @@ defmodule EventStreamex.Operators.Queue do
   end
 
   @doc """
-  Tells the queue that the current item is finished.
+  Tells the queue that the module in the current item has finished.
+
+  It will update its completion state, and if all modules have finished
+  the task is completed
   """
-  @doc since: "1.0.0"
-  def task_finished() do
+  @doc since: "1.2.0"
+  def task_finished(module) do
     Agent.update(__MODULE__, fn queue ->
       case :queue.out(queue) do
         {:empty, new_queue} ->
           new_queue
 
-        {{:value, item}, new_queue} ->
-          {:ok, _res} = QueueStorageAdapter.delete_item(item)
-          new_queue
+        {{:value, {id, {modules, event}} = item}, new_queue} ->
+          new_modules_status =
+            modules
+            |> Map.put(module, true)
+
+          if Enum.all?(new_modules_status |> Map.values()) do
+            {:ok, _res} = QueueStorageAdapter.delete_item(item)
+            new_queue
+          else
+            new_item = {id, {new_modules_status, event}}
+            {:ok, _res} = QueueStorageAdapter.update_processors_status(new_item)
+            :queue.in_r(new_item, new_queue)
+          end
       end
     end)
   end
